backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.4
psycopg[binary]==3.2.6
python-dotenv==1.1.0
Pillow==11.1.0
boto3==1.35.0
gunicorn==22.0.0
celery[redis]==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
firebase-admin==6.6.0
pytz==2024.1
